Upgrade to Play 2.6 - Update dependencies for Play 2.6 and Scala 2.12 - Update config readers to use `get[T]` instead of type specific readers - Refactor `JwtAuthenticatedActions` to `AtlassianHostUserAction` and `OptionalAtlassianHostUserAction` - Update `JwtSignatureCalculator` to use ws dependencies from shaded package - Change config key `play.crypto.sceret` to new key `play.http.secret.key` - Fix all Tests
