+++ modules/core/conf/reference.conf
@@ -13,5 +13,5 @@ addon {
 }
 
 play {
-  crypto.secret = "changeme"
+  http.secret.key  = "changeme"
 }
