+++ project/Dependencies.scala
@@ -20,8 +20,9 @@ object Version {
   val atlassianJwt = "0.1.0"
   val cats = "0.9.0"
   val circe = "0.8.0"
+  val playCirce = "2608.4"
   val scalaUri = "0.4.16"
-  val scalaTestPlusPlay = "2.0.1"
+  val scalaTestPlusPlay = "3.1.1"
   val scalaMock = "3.6.0"
   val scalaCheck = "1.13.5"
 }
@@ -33,7 +34,7 @@ object Library {
   val circeCore = "io.circe" %% "circe-core" % Version.circe
   val circeGeneric = "io.circe" %% "circe-generic" % Version.circe
   val circeParser = "io.circe" %% "circe-parser" % Version.circe
-  val circePlay = "play-circe" %% "play-circe" % s"2.5-${Version.circe}"
+  val circePlay = "play-circe" %% "play-circe" % Version.playCirce
   val scalaUri = "com.netaporter" %% "scala-uri" % Version.scalaUri
   val scalaTestPlusPlay = "org.scalatestplus.play" %% "scalatestplus-play" % Version.scalaTestPlusPlay
   val scalaMock = "org.scalamock" %% "scalamock-scalatest-support" % Version.scalaMock
