+++ project/plugins.sbt
@@ -1,4 +1,4 @@
-addSbtPlugin("com.typesafe.play" % "sbt-plugin" % "2.5.16")
+addSbtPlugin("com.typesafe.play" % "sbt-plugin" % "2.6.3")
 addSbtPlugin("org.scoverage" % "sbt-scoverage" % "1.5.0")
 addSbtPlugin("com.geirsson" % "sbt-scalafmt" % "0.6.8")
 addSbtPlugin("com.github.gseitz" % "sbt-release" % "1.0.5")
